BotFILES/requirements_mcp.txt
# MCP Integration Requirements

# Discord bot
discord.py>=2.3.0

# Web framework for MCP bridge
Flask>=3.0.0
flask-cors>=4.0.0

# HTTP requests for testing
requests>=2.31.0

# JSON handling (built-in, listed for reference)
# json

# Async support (built-in, listed for reference)
# asyncio
